requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
defusedxml>=0.7.1
lxml>=5.0.0
